backend/requirements.txt
Django==1.8.3
djangorestframework==3.1.3
lxml==3.4.4
requests==2.7.0
wheel==0.24.0
